gmail-ai-digest/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain-core
langchain-openai
openai
python-dotenv